+++ src/ArduinoJson/Object/MemberProxy.hpp
@@ -187,8 +187,8 @@ class MemberProxy : public VariantOperators<MemberProxy<TObject, TStringRef> >,
     return _object.getOrAddMember(_key);
   }
 
-  friend bool convertToJson(const this_type &src, VariantRef dst) {
-    return dst.set(src.getUpstreamMember());
+  friend void convertToJson(const this_type &src, VariantRef dst) {
+    dst.set(src.getUpstreamMember());
   }
 
   TObject _object;